[activemq58] Proper prefix on service check
I hacked the StatsDReporter to send the status of activemq.can_connect
instead of activemq_58.can_connect.

It's a bit better than a simple Hack. We decided that the prefix name
shouldn't contain any number, underscore or capital letter (should we
add the dash to this list ?). This way we can have version dependant
configuration in the .yaml config files for JMXFetch without having
multiple and dirty metrics and service check names in the HQ.

UNIT TEST SUITE UPDATE
----------------------

Added a simple unit test to the prefix formatter, when adding versionned
integration like activemq_58, it would be good to check that its prefix
is correctly formatted here.