Limit MBeans query to certain scopes
JMXFetch is currently fetching all available beans regardless of the
user configuration. This might cause a high memory load when the
application is exposing a lot of metrics (thus OOM issues).

This change is the first step towards a bigger JMXFetch memory
consumption improvement: the idea is to find, among the user defined
beans, some common patterns to limit the scopes to query.

For instance,
```
- org.datadog.jmxfetch.test:scope=sameScope,param=sameParam,type=sameType
- org.datadog.jmxfetch.test:scope=sameScope,param=notTheSameParam,type=sameType
```
would induce a query on "org.datadog.jmxfetch.test:scope=sameScope,type=sameType,*"
instead of "*:*".

At the moment, only one scope per domain name is computed: it's kind of
the greatest common divisor/pattern of the beans.

Ideally, this should be improved to recognize multiple scopes among the
same domain name.