Remove host name from service checks payloads
The host name that was sent was the `host` defined in the yaml config for each
instance, which was preventing the backend from correctly assigning host tags
to the service checks.

We remove it completely, the forwarder will take care of adding the correct hostname.

Also, removed unused variables in `TestApp`.