Replace host metadata in dogstasd payload with jmx_server tag
The host name that's sent is the host that's defined in the yaml config
for each instance, which is preventing the backend from correctly assigning
host tags to the service checks.

We fix that by sending a `jmx_server` tag in the dogstatsd payload
instead of a host_name metadata field. The agent's dogstatsd will then
add the agent's hostname.

Also removed unused variables in `TestApp`.