[sc] Split status and service check status
We were sending a service check warning for integrations with too many
metrics which is not the behaviour we'd like to have. Even though there
are too many metrics, the `can_connect` service check should be green.

This was a good oppurtunity for a slight refactoring of `App.java` where
service checks and statuses in the `datadog-agent info` page have been
made completely independant.