[service_discovery] Fix race condition preventing SD initialization
If JMXFetch starts before the collector process creates the SD pipe,
JMXFetch doesn't find the SD pipe and never re-tries opening it, (until
it's fully restarted).

To solve this, retry opening the SD pipe at every run iteration when
SD is enabled. Also, add an `--sd-enabled` option to know whether
JMXFetch should enable SD (i.e. whether it should expect the SD pipe
to exist or not).

The `--sd-standby` option is not used and has therefore been removed.

This change requires a change to `dd-agent` to make `jmxfetch.py` pass
the `--sd-enabled` option when SD-JMXFetch is enabled.