[concurrency] add an unbounded pool thread (seldom-used) for recovery tasks.
We don't want to use an unbounded thread pool during metric collection
to avoid driving CPU usage up. However during the rare occurrences
when we're recovering, using a cachedThreadPool - that will essentially
use a thread per recovering/new instance - makes sense as it's a rare event
and will guarantee equal treatment for all instances. The
cachedThreadPool does not consume resources when idle.

Additionally fix instance cleanup, as that operation results in
an attempt to close the network connection gracefully, but results
in a lock-up when the target pods/ips are no longer up.