[jmxfetch] concurrent metric collection (#199)
* Use ServiceExecutor to fetch metrics within deadline with some concurrency

* [concurrency] add an unbounded pool thread (seldom-used) for recovery tasks.

We don't want to use an unbounded thread pool during metric collection
to avoid driving CPU usage up. However during the rare occurrences
when we're recovering, using a cachedThreadPool - that will essentially
use a thread per recovering/new instance - makes sense as it's a rare event
and will guarantee equal treatment for all instances. The
cachedThreadPool does not consume resources when idle.

Additionally fix instance cleanup, as that operation results in
an attempt to close the network connection gracefully, but results
in a lock-up when the target pods/ips are no longer up.

* fix broken instances management and async cleanup

* Adding a shutdown hook for App related stuff.

* replace ExecutorService if current one hogged-up

* add AppShutdownHook class for App cleanup on shutdown

* do not evict instances we're not certain are broken

* randomize broken instance fix to address potential starvation

* (re)initialize concurrently + some refactor

* configurable threadpools + logging + instance+status pairs

* [tasks] moving to its own package + adding tests for task processors

* fixing service check reporting + tests

* [checkstyle] fixing stylistic issues

* [tasks] TaskProcessor can initialize its own logger