Enable launch of jmxfetch app's executors as daemon, default to NO-daemon
When embedding the jmxfetch app in dd-trce-java we noticed that even if
we it in a thread marked as daemon the jmxfetch app was not terminating
when the main method exited.
After some investigation we noticed that threads in tasks executors were
keeping the thread alive.
This PR DOES NOT change the default behavior, were executors are run as
non-daemons. Adds the possibility, though, to be configured to be
executed as daemon.