Use the name if provided to identify the instance in toString()
When a JMX check is running on an `address:port` and no other
identifiers is provided, the `address:port` is used as an identifier.

However, it could not be unique because we want to be able to run
multiple JMX instances on the same JVM.

This PR modify the `Instance.toString()` method to use the name if one
has been provided in the instance configuration.