Replace LinkedHashMap by HashMap or Map
Configurations come mostly as LinkedHashMap from Yaml or Json parsers
But usage inside JMXFetch is mostly as HashMap.
As LinkedHashMap is heavier regarding memory footprint, we prefer to
avoid its usage by just replacing by HashMap. Also it does not seem
That order matters in the case of JMXFetch usage.
I have changed also the type to use Map interface as it will be easier
To change the implementation if required in the future without
changing this type everywhere.