Cache Metric instances (#273)
* Replace Map metric to a Metric class

During collection process, metrics are collected as
Map<String, Object>, but keys are known and fixed so we can replace
this map to a POJO which will be lighter in term of memory footprint
and allocation.

* cache Metric instances

Use Metric class as a cache and avoid recreating it for each
collection. Only the value is updated. For simple & complex attribute
Tags are also cached into the Metric instance.
For complex & tabular attribute, cache is done per sub attribute into
a new JMXSubAttribute abstract class
Transfer checkName into the attribute to be inserted into the Metric
instance at creation

* fix bad metric instantiantion

Co-authored-by: Jaime Fullaondo <truthbk@gmail.com>