Statsd reporter to require host name and optional port to support UDS (#335)
This change enables JMXFetch to use UDS as well as UDP connections for statsd.
I am introducing a breaking change to the statsd connection string. Now it will require a hostname whereas before the host was optional and it only required a port. Now the host is require and the port is optional. If no port number is found, JMXFetch assumes UDS.

I also rev'd the java-dogstatsd-client version