Avoid thread leak and data race in cleanUpAsync
A thread started in cleanUpAsync may overlap with Instance#init called
from fixBrokenInstances, closing a new freshly created connection
object instead of the old one. This is fixed by capturing specific
connection object in the thread, not the whole instance.

If a connection issue persists over multiple collection intervals,
there will be multiple calls to cleanUpAsync while an Instance
references the same connection object. closeConnector may block to
wait for a network protocol timeout; and, because the implementation
is synchronized, multiple concurrent calls will wait for it
sequentially, causing build-up of threads all waiting to close a
connection. Reset the Instance connection to null, to make sure only
one thread is trying to close a connection.